Fixed Some Errors In Admin Page and Vendor Page Under Product Management

+++ app/Http/Controllers/Backend/ProductController.php
@@ -96,7 +96,7 @@ class ProductController extends Controller
         $categories = Category::latest()->get();
         $subcategory = SubCategory::latest()->get();
         $products = Product::findOrFail($id);
-        $multipleImages = MultiImage::where('product_id', $id)->get();
+        $multipleImages = MultiImage::where('product_id', $id)->latest()->get();
         return view('backend.product.product_edit', compact('activeVendor', 'brands', 'categories', 'subcategory', 'products', 'multipleImages'));
     } //End Method
 
@@ -166,9 +166,52 @@ class ProductController extends Controller
         return redirect()->back()->with($notification);
     } //End Method
 
+    public function AddNewProductMultipleImages(Request $request)
+    {
+        $product_id = $request->id;
+        $images = $request->add_new_multiple_image;
+        if ($images == NULL) {
+            $notification = array(
+                'message' => "Upload Failed Because You Didn't Choose An Image!",
+                'alert-type' => 'error',
+            );
+            return redirect()->back()->with($notification);
+        } else {
+            $request->validate([
+                'add_new_multiple_image.*' => 'image'
+            ], [
+                'add_new_multiple_image.*.image' => 'The uploaded file must be an image in one of the following formats: jpg, jpeg, png, bmp, gif, svg, or webp.',
+            ]);
+            $images = $request->file('add_new_multiple_image');
+            foreach ($images as $image) {
+                $make_name = hexdec(uniqid()) . '_product' . '.' . $image->getClientOriginalExtension();
+                Image::make($image)->resize(1000, 1000)->save('upload/products/multiple_images/' . $make_name);
+                $uploadPath = 'upload/products/multiple_images/' . $make_name;
+
+                MultiImage::insert([
+                    'product_id' => $product_id,
+                    'photo_name' => $uploadPath,
+                    'created_at' => Carbon::now(),
+                ]);
+            }
+            $notification = array(
+                'message' => 'Product Multiple Images Added Successfully!',
+                'alert-type' => 'success',
+            );
+            return redirect()->back()->with($notification);
+        }
+    } //End Method
+
     public function UpdateProductMultipleImages(Request $request)
     {
         $images = $request->multiple_image;
+        if ($images == NULL) {
+            $notification = array(
+                'message' => "Upload Failed Because You Didn't Choose An Image!",
+                'alert-type' => 'error',
+            );
+            return redirect()->back()->with($notification);
+        }
         foreach ($images as $id => $image) {
             $imageDelete = MultiImage::findOrFail($id);
             unlink($imageDelete->photo_name);
